Move engine-only query tests
